+++ astroid/inference_tip.py
@@ -9,7 +9,7 @@ import typing
 import wrapt
 
 from astroid import bases, util
-from astroid.cache import cached_generator
+from astroid.cache import LRUCache
 from astroid.exceptions import InferenceOverwriteError, UseInferenceDefault
 from astroid.nodes import NodeNG
 from astroid.typing import InferFn
@@ -18,32 +18,28 @@ InferOptions = typing.Union[
     NodeNG, bases.Instance, bases.UnboundMethod, typing.Type[util.Uninferable]
 ]
 
-_cache: typing.Dict[
-    typing.Tuple[InferFn, NodeNG], typing.Optional[typing.List[InferOptions]]
-] = {}
 
-
-def clear_inference_tip_cache():
-    """Clear the inference tips cache."""
-    _cache.clear()
+_INFERENCE_TIP_CACHE: LRUCache = LRUCache()
 
 
 @wrapt.decorator
-def _inference_tip_cached(
-    func: InferFn, instance: None, args: typing.Any, kwargs: typing.Any
-) -> typing.Iterator[InferOptions]:
-    """Cache decorator used for inference tips"""
-    node = args[0]
-    try:
-        result = _cache[func, node]
-        # If through recursion we end up trying to infer the same
-        # func + node we raise here.
+def _cached_generator(
+    func: InferFn,
+    instance: typing.Any,
+    args: typing.Tuple[typing.Any, ...],
+    kwargs: typing.Dict[str, typing.Any],
+) -> typing.Any:
+    key = func, args[0]
+
+    if key in _INFERENCE_TIP_CACHE:
+        result = _INFERENCE_TIP_CACHE[key]
+
         if result is None:
             raise UseInferenceDefault()
-    except KeyError:
-        _cache[func, node] = None
-        result = _cache[func, node] = list(func(*args, **kwargs))
-        assert result
+    else:
+        _INFERENCE_TIP_CACHE[key] = None
+        result = _INFERENCE_TIP_CACHE[key] = list(func(*args, **kwargs))
+
     return iter(result)
 
 
@@ -85,7 +81,7 @@ def inference_tip(infer_function: InferFn, raise_on_overwrite: bool = False) -> 
             )
 
         # pylint: disable=no-value-for-parameter
-        node._explicit_inference = cached_generator()(infer_function)
+        node._explicit_inference = _cached_generator(infer_function)
         return node
 
     return transform
